raphael: introduce gfscreenoffd
This is a small daemon to capture fingerprint touch event while screen is off.
goodix_ts fire a specific input event when the area of fingerprint
is being touched on screen off but fod_status is on.
This daemon will create an another uinput and act as a virtual input
based on goodix_ts's specific screen off event.

Inspired from Okitavera's FOD module for Mi 9 SE and TheScarastic's
onPress() IFingerprintInscreenCallback in reversed SystemUI.

Signed-off-by: Nauval Rizky <enuma.alrizky@gmail.com>