raphael: anxfilecheck: Run script only at boot time
When it's not marked as `disabled` and `oneshot`, this script will be called
again and again by init, which is wasting resources.

Mark it properly so it only check and prepare the anx files only when
it's called; which is after boot completed.
And additionally, add method for waiting the sdcard for write-ready
before copying the files.

Signed-off-by: Nauval Rizky <enuma.alrizky@gmail.com>