raphael: Remove ANX-filecheck scripts
With this changes (and of course, proper selinux implementation on
device/{aosip,qcom/sepolicy), we should be able to build rom in user
variant.

For these ANX situation, we have another way to mitigate it.
- Rely on Magisk's version for ANX Camera
- Use modified ANX Camera that has system-wide FeatureParser, which
  parse the config from /system/etc/

For the 2nd solution, you can use
https://github.com/nullxception/vendor_MiuiCamera as your main anxcamera
repository, which based on https://gitlab.com/nullxception/anxcamera-custom

Signed-off-by: Nauval Rizky <enuma.alrizky@gmail.com>